# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for managing user profiles, authentication, analytics, and more.",
        "title": "Cash Advisor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Mikhail Vakhrushin",
            "email": "wachrusz@gmail.com"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/profile/get": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get the user profile for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error getting user profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/profile/update-name": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update the user profile for the authenticated user with a new name and surname.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile with name",
                "parameters": [
                    {
                        "description": "New name to be added to the profile",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error updating user profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "JWT": {
            "description": "To authorize,",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}